frontend: Split ApartmentNewSalePage components to separate files
