+++ frontend/src/features/apartment/ApartmentNewSalePage/MaximumPriceModalError.tsx
@@ -1,0 +1,25 @@
+import {FetchBaseQueryError} from "@reduxjs/toolkit/query";
+import {Button, Dialog} from "hds-react";
+
+const MaximumPriceModalError = ({error, setIsModalVisible}) => {
+    const nonFieldError = ((error as FetchBaseQueryError)?.data as {message?: string})?.message || "";
+    return (
+        <>
+            <Dialog.Content>
+                <p>Virhe: {(error as FetchBaseQueryError)?.status}</p>
+                <p>{nonFieldError}</p>
+            </Dialog.Content>
+            <Dialog.ActionButtons>
+                <Button
+                    onClick={() => setIsModalVisible(false)}
+                    variant="secondary"
+                    theme="black"
+                >
+                    Sulje
+                </Button>
+            </Dialog.ActionButtons>
+        </>
+    );
+};
+
+export default MaximumPriceModalError;
